Dedicated ModPresence class
