+++ src/main/java/net/joseph/vaultfilters/VaultFilters.java
@@ -19,16 +19,10 @@ import net.joseph.vaultfilters.attributes.soul.*;
 import net.joseph.vaultfilters.attributes.tool.ToolMaterialAttribute;
 import net.joseph.vaultfilters.attributes.trinket.*;
 import net.joseph.vaultfilters.configs.VFServerConfig;
-import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.MinecraftServer;
-import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.level.Level;
-import net.minecraftforge.api.distmarker.Dist;
-import net.minecraftforge.api.distmarker.OnlyIn;
-import net.minecraftforge.client.event.ClientPlayerNetworkEvent;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.event.TickEvent;
-import net.minecraftforge.event.entity.player.PlayerEvent;
 import net.minecraftforge.event.world.WorldEvent;
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
@@ -37,14 +31,7 @@ import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.config.ModConfig;
 import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
-import net.minecraftforge.network.NetworkRegistry;
-import net.minecraftforge.network.PacketDistributor;
-import net.minecraftforge.network.simple.SimpleChannel;
 import org.slf4j.Logger;
-
-import java.util.HashSet;
-import java.util.Set;
-import java.util.UUID;
 
 //import com.simibubi.create.content.logistics.filter.FilterItemStack;
 
@@ -57,21 +44,12 @@ public class VaultFilters {
     public static final int NO_CACHE_FLAG = 457;
     public static final Logger LOGGER = LogUtils.getLogger();
 
-    public static final Set<UUID> PLAYERS_WITH_VAULT_FILTERS = new HashSet<>();
-    public static boolean serverHasVaultFilters = false;
-    private static final String PROTOCOL_VERSION = "1";
-    public static final SimpleChannel CHANNEL = NetworkRegistry.newSimpleChannel(
-            new ResourceLocation(MOD_ID, "main"),
-            () -> PROTOCOL_VERSION,
-            o -> true,
-            o -> true
-    );
-
     public VaultFilters() {
         IEventBus eventBus = FMLJavaModLoadingContext.get().getModEventBus();
         eventBus.addListener(this::setup);
         //eventBus.addListener(this::onWorldLoad);
         MinecraftForge.EVENT_BUS.register(this);
+        MinecraftForge.EVENT_BUS.register(ModPresence.class);
     }
 
     public static Level LEVEL_REF;
@@ -93,26 +71,18 @@ public class VaultFilters {
             if (ticks >= 60*20) { //60 seconds * 20 tps
                 ticks = 0;
                 //VaultFilters.LOGGER.info("cleared");
-                VaultFilters.LOGGER.info("Clearing Vault Filters cache with " + VFCache.cacheMap.size() + "elements");
                 VFCache.cacheMap.forEach((key, value) -> {
                     if (value.TTK == 0) {
                         VFCache.cacheMap.remove(key);
                     }
                     value.TTK--;
                 });
-                VaultFilters.LOGGER.info("Vault Filters cache cleared " + VFCache.cacheMap.size() + "elements remaining");
             }
         }
     }
 
     private void setup(FMLCommonSetupEvent event) {
-        // Mod Detection System
-        CHANNEL.messageBuilder(ModPresenceMessage.class, 0)
-                .encoder(ModPresenceMessage::encoder)
-                .decoder(ModPresenceMessage::decoder)
-                .consumer(ModPresenceMessage::consumer)
-                .add();
-
+        ModPresence.init();
         ModLoadingContext.get().registerConfig(ModConfig.Type.SERVER, VFServerConfig.SPEC, "vaultfilters-server.toml");
         // This has a specific order as this controls the order displayed in the filters
         new ItemNameAttribute(("Vault Helmet")).register(ItemNameAttribute::new);
@@ -192,29 +162,5 @@ public class VaultFilters {
         new CatalystModifierCategoryAttribute("Bonus Chests").register(CatalystModifierCategoryAttribute::new);
 
 
-    }
-
-    @OnlyIn(Dist.CLIENT) @SubscribeEvent
-    public void onClientConnect(ClientPlayerNetworkEvent.LoggedInEvent event) {
-        CHANNEL.sendToServer(new ModPresenceMessage(MOD_VERSION));
-    }
-
-    @OnlyIn(Dist.CLIENT) @SubscribeEvent
-    public void onClientDisconnect(ClientPlayerNetworkEvent.LoggedOutEvent event) {
-        serverHasVaultFilters = false;
-    }
-
-    @OnlyIn(Dist.DEDICATED_SERVER) @SubscribeEvent
-    public void onPlayerLoggedIn(PlayerEvent.PlayerLoggedInEvent event) {
-        if (event.getPlayer() instanceof ServerPlayer player) {
-            CHANNEL.send(PacketDistributor.PLAYER.with(() -> player), new ModPresenceMessage(MOD_VERSION));
-        }
-    }
-
-    @OnlyIn(Dist.DEDICATED_SERVER) @SubscribeEvent
-    public void onPlayerLoggedOut(PlayerEvent.PlayerLoggedOutEvent event) {
-        if (event.getPlayer() instanceof ServerPlayer player) {
-            PLAYERS_WITH_VAULT_FILTERS.remove(player.getUUID());
-        }
     }
 }
