+++ src/main/java/net/joseph/vaultfilters/mixin/compat/create/MixinAttributeFilterScreen.java
@@ -2,7 +2,7 @@ package net.joseph.vaultfilters.mixin.compat.create;
 
 import com.simibubi.create.content.logistics.filter.AttributeFilterScreen;
 import com.simibubi.create.content.logistics.filter.ItemAttribute;
-import net.joseph.vaultfilters.VaultFilters;
+import net.joseph.vaultfilters.ModPresence;
 import net.joseph.vaultfilters.attributes.abstracts.VaultAttribute;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
@@ -16,7 +16,7 @@ import java.util.List;
 public class MixinAttributeFilterScreen {
     @Redirect(at = @At(value = "INVOKE", target = "Lcom/simibubi/create/content/logistics/filter/ItemAttribute;listAttributesOf(Lnet/minecraft/world/item/ItemStack;Lnet/minecraft/world/level/Level;)Ljava/util/List;"), method = "referenceItemChanged", remap = false)
     public List<ItemAttribute> listAttributes(ItemAttribute instance, ItemStack stack, Level world) {
-        if (!VaultFilters.serverHasVaultFilters && instance instanceof VaultAttribute<?>) {
+        if (!ModPresence.serverHasVaultFilters && instance instanceof VaultAttribute<?>) {
             return List.of();
         }
         return instance.listAttributesOf(stack, world);
